requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.4
logging==0.4.9.6
mysqlclient==2.2.0
numpy==1.24.4
pandas==2.0.3
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.1.0
